Group text examples on styleguide (#6800)
* Group `Text` heading examples

* Group `Text` body examples

* Group `Text` align examples

* Group `font-weight` examples

* Group `color` text examples

* Update polaris.shopify.com/content/components/text.md


* Update polaris.shopify.com/content/components/text.md


* Update text color examples

* Update text weight examples order
