+++ polaris.shopify.com/content/components/text.md
@@ -22,98 +22,26 @@ status:
   value: Alpha
   message: This component is in development. There could be breaking changes made to it in a non-major release of Polaris. Please use with caution.
 examples:
-  - fileName: text-heading-small.tsx
-    title: Heading small
+  - fileName: text-heading.tsx
+    title: Heading
     description: >-
-      Use to create a small heading text.
-  - fileName: text-heading-medium.tsx
-    title: Heading medium
+      Use to create various levels of hierarchy on the page.
+  - fileName: text-body.tsx
+    title: Body
     description: >-
-      Use to create a medium heading text.
-  - fileName: text-heading-large.tsx
-    title: Heading large
+      Use to create a range of body text. These styles are primarily used within components and blocks of text.
+  - fileName: text-align.tsx
+    title: Align
     description: >-
-      Use to create a large heading text.
-  - fileName: text-heading-extra-large.tsx
-    title: Heading extra large
+      Use to set text alignment.
+  - fileName: text-weight.tsx
+    title: Weight
     description: >-
-      Use to create an extra large heading text.
-  - fileName: text-heading-2xl.tsx
-    title: Heading 2xl
+      Use to give text a range of font weights.
+  - fileName: text-color.tsx
+    title: Color
     description: >-
-      Use to create a 2xl heading text.
-  - fileName: text-heading-3xl.tsx
-    title: Heading 3xl
-    description: >-
-      Use to create a 3xl heading text.
-  - fileName: text-heading-4xl.tsx
-    title: Heading 4xl
-    description: >-
-      Use to create a 4xl heading text.
-  - fileName: text-body-small.tsx
-    title: Body small
-    description: >-
-      Use to create a small body text.
-  - fileName: text-body-medium.tsx
-    title: Body medium
-    description: >-
-      Use to create a medium body text.
-  - fileName: text-body-large.tsx
-    title: Body large
-    description: >-
-      Use to create a large body text.
-  - fileName: text-with-align-inherit.tsx
-    title: With align inherit
-    description: >-
-      Use to inherit parent alignment.
-  - fileName: text-with-align-start.tsx
-    title: With align start
-    description: >-
-      Use to align text at start of horizontal line.
-  - fileName: text-with-align-center.tsx
-    title: With align center
-    description: >-
-      Use to align text at center of horizontal line.
-  - fileName: text-with-align-end.tsx
-    title: With align end
-    description: >-
-      Use to align text at end of horizontal line.
-  - fileName: text-with-align-justify.tsx
-    title: With align justify
-    description: >-
-      Use to align text on a vertical line.
-  - fileName: text-with-font-weight-regular.tsx
-    title: With font weight regular
-    description: >-
-      Use to give text a regular font weight.
-  - fileName: text-with-font-weight-medium.tsx
-    title: With font weight medium
-    description: >-
-      Use to give text a medium font weight.
-  - fileName: text-with-font-weight-semibold.tsx
-    title: With font weight semibold
-    description: >-
-      Use to give text a semi-bold font weight.
-  - fileName: text-with-font-weight-bold.tsx
-    title: With font weight bold
-    description: >-
-      Use to give text a bold font weight.
-  - fileName: text-with-color-success.tsx
-    title: With color success
-    description: >-
-      Use in combination with a symbol showing an increasing value to indicate an upward trend.
-  - fileName: text-with-color-critical.tsx
-    title: With color critical
-    description: >-
-      Use in combination with a symbol showing a decreasing value to indicate a downward trend.
-  - fileName: text-with-color-warning.tsx
-    title: With color warning
-    description: >-
-      Use to denote something that needs attention, or that merchants need to take action on.
-  - fileName: text-with-color-subdued.tsx
-    title: With color subdued
-    description: >-
-      Use to de-emphasize a piece of text that is less important to merchants than other nearby text. May also be used to indicate when normal content is absent, for example, “No supplier listed”. Don’t use only for aesthetic effect.
+      Use to set text color.
 ---
 
 # Text
