--- polaris.shopify.com/src/pages/examples/text-with-font-weight-regular.tsx
@@ -1,13 +1,0 @@
-import { Text } from "@shopify/polaris";
-import React from "react";
-import { withPolarisExample } from "../../components/PolarisExamplePage";
-
-function TextExample() {
-  return (
-    <Text variant="bodyMd" as="p" fontWeight="regular">
-      Sales this year
-    </Text>
-  );
-}
-
-export default withPolarisExample(TextExample);
